Renamed config consolidation and added functionality
