+++ Utils/backups_json.json
@@ -1,7 +1,7 @@
 {
     "Macro_Consolidation": {
-        "filename": "Macro_Consolidation_Template.zip",
-        "checksum": "2e34df599076bd44d62b9a2b80adcde6ba3b649a4b0010fbea074ef974a406e63b80929d9c9f1d311a5be942f6cf164ad8d2f1f91de08cee09fb1e0c11047bb8"
+        "filename": "Macro_Consolidation_TemplateV3.zip",
+        "checksum": "cfbec54620df3d5b5f1cfd9d082255b538834b531eb882c1762716e7325138af08ab3b0fbc1fa214455ee5627a876d9259dff979d526663cddafac28ad7a2bad"
     },
     "OS11-20_Config_Update": {
         "filename": "OS1120_Update_Template.zip",
